+++ sphinx/source/changelog.rst
@@ -4,6 +4,8 @@ Changelog
 10.0b1
 ------
 * Dropped support for Python 3.9
+* Added the ``--ay`` option to :ref:`trace.py` (for capturing AY audio when
+  writing a WAV file)
 * Added support for the MEMPTR register to CMIOSimulator and CCMIOSimulator
 * Added the ``memptr`` parameter to the :ref:`SIM` macro (to set the MEMPTR
   register)